fix(mcp): split file-upload helpers into _fileupload; harden base64 cap
CI module-size ratchet (ADR-0008): source_upload_bytes pushed
mcp/tools/sources.py to 1103 lines, over the 1000-line budget. Extract the
file-transfer slice — _broker_upload (signed-URL), _decode_upload_b64 +
_add_bytes (in-channel byte upload), and the shared _add_one plan/execute seam —
into a sibling tools/_fileupload.py (the established drain-below-budget pattern).
sources.py is back to 900 lines.

Also addresses the gemini-code-assist review: the base64 cap now runs a
fast-fail length pre-check BEFORE the whitespace strip (bounds allocation), then
checks the WHITESPACE-STRIPPED length against the cap. This fixes a real edge
bug — the old raw-length check wrongly rejected a valid 76-col-wrapped payload
near the cap, because its newlines pushed the raw length over the limit even
though the decoded content was within it. New regression test locks it in.

- No behavior change to the tool contract, params, cap value, or error strings
  tests assert on (the too-large message still names the source_add fallback).
- Test constant reference repointed to _fileupload; +1 wrapped-near-cap test.
